Finished implementing other verbs.
