--- src/api/base/types.ts
@@ -1,8 +1,0 @@
-import { NextFunction, Response, Request } from 'express';
-
-export type ApiHttpHandler = (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-) => Promise<Response<unknown> | void>;
-export type ApiResponse = Promise<Response<unknown> | void>;
